deleted unused jsx files and separated the router files from the inner screen files
