+++ src/client/inners/detail_screen.jsx
